+++ benchmarks/hot_path_latency/quill_hot_path_rdtsc_clock.cpp
@@ -56,7 +56,8 @@ void quill_benchmark(std::vector<uint16_t> const& thread_count_array,
 
   Logger* logger = Frontend::create_or_get_logger(
     "bench_logger", std::move(file_sink),
-    "%(time) [%(thread_id)] %(short_source_location) %(log_level) %(message)");
+    "%(time) [%(thread_id)] %(short_source_location) %(log_level) %(message)", "%H:%M:%S.%Qns",
+    quill::Timezone::LocalTime, false);
 
   /** LOGGING THREAD FUNCTIONS - on_start, on_exit, log_func must be implemented **/
   /** those run on a several thread(s). It can be one or multiple threads based on THREAD_LIST_COUNT config */